Split api.py into HTTP client and pure parser module
Extract all stateless HTML/text parsing functions into api_parser.py
so they can be tested and imported without aiohttp or auth context.
api.py now contains only network I/O and delegates parsing via imports.
Update tests to import parser functions directly from api_parser.
